Create OOP_WaterJug Solution
Problem Statement:

You are given two jugs, a 4-gallon one and a 3-gallon one, with no measuring marks on them. You also have access to a pump that can fill the jugs with water. The goal is to measure exactly 2 gallons of water using these jugs. You can perform the following actions:

Fill the 4-gallon jug.
Fill the 3-gallon jug.
Empty the 4-gallon jug.
Empty the 3-gallon jug.
Pour water from the 3-gallon jug into the 4-gallon jug until the 4-gallon jug is full.
Pour water from the 4-gallon jug into the 3-gallon jug until the 3-gallon jug is full.
Pour all the water from the 3-gallon jug into the 4-gallon jug.
Pour all the water from the 4-gallon jug into the 3-gallon jug.
We need to find a sequence of actions to achieve the goal of having exactly 2 gallons of water in the 4-gallon jug while ensuring the 3-gallon jug is empty.